LangGraph-tutorial/pyproject.toml
[project]
name = "langgraph-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain[anthropic,antrhopic,openai]>=0.3.25",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
]
